backend/requirements.txt
fastapi
uvicorn
pydantic
bcrypt==4.0.1
boto3>=1.38.13
fastapi>=0.115.12
mangum>=0.19.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.4.0
python-multipart>=0.0.20
uvicorn>=0.34.2
awslambdaric>=3.0.2
simplejson>=3.20.1
ruff

pytest
pytest-cov
httpx
